requirements.txt
discord.py
nba_api
python-dotenv
pillow
pandas
requests
